Refactor - extract tests for refund reasons
